Split automation heartbeat reviews
